+++ src/ring/swagger/common.clj
@@ -1,12 +1,6 @@
-(ns ring.swagger.common)
-
-(defmacro fn->
-  "Creates a function that threads on input with some->"
-  [& body] `(fn [x#] (some-> x# ~@body)))
-
-(defmacro fn->>
-  "Creates a function that threads on input with some->>"
-  [& body] `(fn [x#] (some->> x# ~@body)))
+(ns ring.swagger.common
+  (:require [plumbing.core :refer [fn->]]
+            [camel-snake-kebab :as csk]))
 
 (defn remove-empty-keys
   "removes empty keys from a map"
@@ -48,3 +42,6 @@
             form       (drop (* 2 (count parameters)) c)]
         [parameters form])
       [{} c])))
+
+(defn ->CamelCase [x]
+  (csk/->CamelCase x))