+++ test/ring/swagger/schema_test.clj
@@ -60,26 +60,6 @@
   (fact "without millis"
     (coerce! {:d Date} {:d "2014-02-24T21:37:40Z"}) =not=> (throws Exception)))
 
-(fact "models"
-
-  (fact "model?"
-    (model? AllTypes) => true
-    (model? 'AllTypes) => false
-    (model? #'AllTypes) => false
-    (model? {:a String}) => false)
-
-  (fact "model-var"
-    (model-var AllTypes) => #'AllTypes
-    (model-var 'AllTypes) => #'AllTypes
-    (model-var #'AllTypes) => #'AllTypes
-    (model-var {:a String}) => nil)
-
-  (fact "model-name"
-    (model-name AllTypes) => "AllTypes"
-    (model-name 'AllTypes) => "AllTypes"
-    (model-name #'AllTypes) => "AllTypes"
-    (model-name {:a String}) => nil))
-
 (facts "types"
 
   (fact "basic types can act as fields"
@@ -107,10 +87,10 @@
     (eval '(defmodel MapModel [String])) => (throws AssertionError))
 
   (fact "has meta-data"
-    MapModel => (has-meta {:model #'MapModel :name 'MapModel}))
+    MapModel => (has-meta {:name 'MapModel}))
   (fact "model?"
-    (model? MapModel) => true
-    (model? {:a String}) => false))
+    (s/schema-name MapModel) => 'MapModel
+    (s/schema-name {:a String}) => falsey))
 
 (fact "field"
 
@@ -127,6 +107,20 @@
 
   (fact "field set meta-data to it"
     (field String {:kikka :kakka}) => (has-meta {:kikka :kakka})))
+
+(fact "named-schema"
+
+  (fact "schema is named"
+    (s/defschema AbbaSchema {:s String})
+    AbbaSchema => named-schema?)
+
+  (fact "model is named"
+    (defmodel AbbaModel {:s String})
+    AbbaModel => named-schema?)
+
+  (fact "def is not named"
+    (def AbbaDef {:s String})
+    AbbaDef =not=> named-schema?))
 
 (fact "coercion"
   (let [valid {:a "kikka"}
@@ -160,13 +154,13 @@
 
 (facts "nested models"
   (fact ".. have generated sub-models and are referenced from parent"
-    (model-var (get-in Customer [:address])) => #'CustomerAddress
+    (s/schema-name (get-in Customer [:address])) => 'CustomerAddress
     CustomerAddress => {:street String
                         (s/optional-key :country) {:code #{(s/enum :fi :sv)}
                                                    :name String}})
 
   (fact ".. for deeper level also"
-    (model-var (get-in Customer [:address (s/optional-key :country)])) => #'CustomerAddressCountry
+    (s/schema-name (get-in Customer [:address (s/optional-key :country)])) => 'CustomerAddressCountry
     CustomerAddressCountry => {:code #{(s/enum :fi :sv)}
                                :name String}))